requirements.txt
langchain==0.3.25
langchain_community==0.3.23
langchain_google_genai==2.1.4
python-dotenv==1.1.0
streamlit==1.44.1